+++ src/GlobalPayments.Api/Builders/ManagementBuilder.cs
@@ -40,6 +40,7 @@ namespace GlobalPayments.Api.Builders {
         internal IEnumerable<DisputeDocument> DisputeDocuments { get; set; }
         internal string DisputeId { get; set; }
         internal string DynamicDescriptor { get; set; }
+        internal eCheck ECheck { get; set; }
         internal decimal? Gratuity { get; set; }
         internal string IdempotencyKey { get; set;  }
         internal string InvoiceNumber { get; set; }
@@ -221,6 +222,20 @@ namespace GlobalPayments.Api.Builders {
         }
 
         /// <summary>
+        /// Set the election check information
+        /// </summary>
+        /// <remarks>
+        /// This value is sent during the authorization process and is displayed
+        /// in the consumer's account.
+        /// </remarks>
+        /// <param name="value">eCheck</param>
+        /// <returns>AuthorizationBuilder</returns>
+        public ManagementBuilder WithBankTransferData(eCheck value) {
+            ECheck = value;
+            return this;
+        }
+
+        /// <summary>
         /// Sets the transaction's dynamic descriptor.
         /// </summary>
         /// <remarks>
@@ -229,7 +244,8 @@ namespace GlobalPayments.Api.Builders {
         /// </remarks>
         /// <param name="value">The dynamic descriptor</param>
         /// <returns>AuthorizationBuilder</returns>
-        public ManagementBuilder WithDynamicDescriptor(string value) {
+        public ManagementBuilder WithDynamicDescriptor(string value)
+        {
             DynamicDescriptor = value;
             return this;
         }
@@ -413,6 +429,9 @@ namespace GlobalPayments.Api.Builders {
         protected override void SetupValidations() {
             Validations.For(TransactionType.Capture | TransactionType.Edit | TransactionType.Hold | TransactionType.Release)
                 .Check(() => PaymentMethod).IsNotNull();
+
+            Validations.For(TransactionType.Capture | TransactionType.Edit | TransactionType.Hold | TransactionType.Release | TransactionType.Reauth)
+                .Check(() => TransactionId).IsNotNull();
 
             // TODO: Need level validations
             //Validations.For(TransactionType.Edit).With(TransactionModifier.Level_II)