+++ src/GlobalPayments.Api/ServicesContainer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using GlobalPayments.Api.Entities;
 using GlobalPayments.Api.Gateways;
@@ -76,7 +77,7 @@ namespace GlobalPayments.Api {
     /// internally by exposed APIs throughout the SDK.
     /// </remarks>
     public class ServicesContainer : IDisposable {
-        private Dictionary<string, ConfiguredServices> _configurations;
+        private ConcurrentDictionary<string, ConfiguredServices> _configurations;
         private static ServicesContainer _instance;
 
         internal static ServicesContainer Instance {
@@ -121,7 +122,7 @@ namespace GlobalPayments.Api {
         }
 
         private ServicesContainer() {
-            _configurations = new Dictionary<string, ConfiguredServices>();
+            _configurations = new ConcurrentDictionary<string, ConfiguredServices>();
         }
 
         private ConfiguredServices GetConfiguration(string configName) {
@@ -133,7 +134,9 @@ namespace GlobalPayments.Api {
         private void AddConfiguration(string configName, ConfiguredServices config) {
             if (_configurations.ContainsKey(configName))
                 _configurations[configName] = config;
-            else _configurations.Add(configName, config);
+            else if(!_configurations.TryAdd(configName, config)) {
+                throw new ConfigurationException($"Failed to add configuration: {configName}.");
+            }
         }
 
         internal IPaymentGateway GetClient(string configName) {
@@ -212,7 +215,10 @@ namespace GlobalPayments.Api {
 
         internal void removeConfiguration(String configName) {
             if(_configurations.ContainsKey(configName)) {
-                _configurations.Remove(configName);
+                ConfiguredServices config = new ConfiguredServices();
+                if(!_configurations.TryRemove(configName,out config)) {
+                    throw new ConfigurationException($"Failed to remove configuration: {configName}.");
+                }
             }
         }