+++ tests/GlobalPayments.Api.Tests/GpApi/GpApiAvsCheckTestCards.cs
@@ -1,0 +1,34 @@
+﻿namespace GlobalPayments.Api.Tests.GpApi
+{
+    public static class GpApiAvsCheckTestCards
+    {
+        public const string AVS_MASTERCARD_1 = "5167308114614170";
+        public const string AVS_MASTERCARD_2 = "5167300431085507";
+        public const string AVS_MASTERCARD_3 = "5167302774869323";
+        public const string AVS_MASTERCARD_4 = "5167308372298823";
+        public const string AVS_MASTERCARD_5 = "5167304376624205";
+        public const string AVS_MASTERCARD_6 = "5167305286858551";
+        public const string AVS_MASTERCARD_7 = "5167302027021516";
+        public const string AVS_MASTERCARD_8 = "5167300747736579";
+        public const string AVS_MASTERCARD_9 = "5167300575629722";
+        public const string AVS_MASTERCARD_10 = "5167305971795258";
+        public const string AVS_MASTERCARD_11 = "5167306563324952";
+        public const string AVS_MASTERCARD_12 = "5167305301636255";
+        public const string AVS_MASTERCARD_13 = "5167308925965308";
+        public const string AVS_MASTERCARD_14 = "5167308304373025";
+        public const string AVS_VISA_1 = "4259917010910078";
+        public const string AVS_VISA_2 = "4259916460134858";
+        public const string AVS_VISA_3 = "4259910583250337";
+        public const string AVS_VISA_4 = "4259913676419876";
+        public const string AVS_VISA_5 = "4259915858684813";
+        public const string AVS_VISA_6 = "4259919754924299";
+        public const string AVS_VISA_7 = "4259917979151326";
+        public const string AVS_VISA_8 = "4259915339571613";
+        public const string AVS_VISA_9 = "4259911964632119";
+        public const string AVS_VISA_10 = "4259912262216019";
+        public const string AVS_VISA_11 = "4259912836281333";
+        public const string AVS_VISA_12 = "4259914539794240";
+        public const string AVS_VISA_13 = "4259913732291939";
+        public const string AVS_VISA_14 = "4259913622000341";
+    }
+}